[optimization]: optimized brightness control
